## Proposed changes
```
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
BM_UInt128HashCRC32SSE      0.770 ns        0.770 ns   1000000000
BM_UInt128HashCRC32ARM      0.417 ns        0.417 ns   1000000000 
```

<!--Describe your changes.-->